+++ libraries/extractor/src/main/java/androidx/media3/extractor/text/SubtitleTranscodingExtractorOutput.java
@@ -46,7 +46,7 @@ public final class SubtitleTranscodingExtractorOutput implements ExtractorOutput
   private final SubtitleParser.Factory subtitleParserFactory;
   private final SparseArray<SubtitleTranscodingTrackOutput> textTrackOutputs;
 
-  private boolean hasNonTextTracks;
+  private boolean hasTracksOtherThanTextAndMetadata;
 
   public SubtitleTranscodingExtractorOutput(
       ExtractorOutput delegate, SubtitleParser.Factory subtitleParserFactory) {
@@ -65,8 +65,10 @@ public final class SubtitleTranscodingExtractorOutput implements ExtractorOutput
 
   @Override
   public TrackOutput track(int id, @C.TrackType int type) {
+    if (type != C.TRACK_TYPE_TEXT && type != C.TRACK_TYPE_METADATA) {
+      hasTracksOtherThanTextAndMetadata = true;
+    }
     if (type != C.TRACK_TYPE_TEXT) {
-      hasNonTextTracks = true;
       return delegate.track(id, type);
     }
     SubtitleTranscodingTrackOutput existingTrackOutput = textTrackOutputs.get(id);
@@ -82,7 +84,7 @@ public final class SubtitleTranscodingExtractorOutput implements ExtractorOutput
   @Override
   public void endTracks() {
     delegate.endTracks();
-    if (hasNonTextTracks) {
+    if (hasTracksOtherThanTextAndMetadata) {
       for (int i = 0; i < textTrackOutputs.size(); i++) {
         textTrackOutputs.valueAt(i).shouldSuppressParsingErrors(true);
       }